Set up a proper post processing stack
This does follow the "simple approach" to post processing. Eventually I might like to convert to the "advanced post processing" tutorial that is in the Godot documentation, especially since that gives the post process effect access to normal and depth information in addition to colour. Which would allow some more advanced effects. I have an idea for depth-based chroma aberrations. But that would absolutely require the advanced setup