It's a Banger!
Debugging is properly setup for VSCode (assuming a GODOT4 environment variable). Also finalized much of the feature set for the item system. Moved that over from the testing setup to a usable setup for games. And got started working on a save station setup, mainly focused on allowing players to create new save slots or overwrite one.