Finished Screen FX; Cleared unneeded files
RetroPBR shader is now screen space, and simplified to make it more efficient. Also cleared out a bunch of extra testing files. I briefly had a minecraft texture pack that I used for testing, but I removed the files before committing to the repo so I'm not taking anyone's stuff. IDK the license on the pack, but I imagine they don't allow taking it for my game. So testing is OK, but nothing else