Footstep System & Working on sky
The footstep system is a component that can be added to any dynamic entity and uses the change in horizontal position and rotation to determine when to play a stepping sound. It gets the step sounds by polling for ground material components on the ground physics objects.
Also working a bit more on the dynamic sky system. In editor the swapping between sun and moon is working well, but it doesn't seem to work as well in game for some reason. I'll have to do some research into why that it