Utility AI
Built a simple structure for using a utility selection. It only provides the most optimal action, it doesn't perform any particular action on its own. It also includes a few helpful systems such as fuzziness to ensure the actions are less robotic and an option to choose no action when all weights are negative (meaning they are really bad choices).